ast_results/hidroh_materialistic/app/src/main/java/io/github/hidroh/materialistic/preference/FontPreference.java
// isComment
package io.github.hidroh.materialistic.preference;

import android.content.Context;
import android.text.TextUtils;
import android.util.AttributeSet;
import android.view.LayoutInflater;
import android.view.View;
import android.view.ViewGroup;
import android.widget.TextView;
import io.github.hidroh.materialistic.Application;
import io.github.hidroh.materialistic.FontCache;
import io.github.hidroh.materialistic.R;

public class isClassOrIsInterface extends SpinnerPreference {

    private final LayoutInflater isVariable;

    @SuppressWarnings("isStringConstant")
    public isConstructor(Context isParameter, AttributeSet isParameter) {
        this(isNameExpr, isNameExpr, isIntegerConstant);
    }

    public isConstructor(Context isParameter, AttributeSet isParameter, int isParameter) {
        super(isNameExpr, isNameExpr, isNameExpr);
        isNameExpr = isNameExpr.isMethod(isMethod());
    }

    @Override
    protected View isMethod(int isParameter, ViewGroup isParameter) {
        return isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr, true);
    }

    @Override
    protected void isMethod(int isParameter, View isParameter) {
        TextView isVariable = (TextView) isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr.isFieldAccessExpr);
        isNameExpr.isMethod(isNameExpr.isMethod().isMethod(isMethod(), isNameExpr[isNameExpr]));
        isNameExpr.isMethod(isNameExpr[isNameExpr]);
    }

    @Override
    protected boolean isMethod(String isParameter) {
        if (isNameExpr.isMethod(isMethod(), isMethod().isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr))) {
            isNameExpr.isFieldAccessExpr = isNameExpr.isMethod().isMethod(isMethod(), isNameExpr);
        }
        return super.isMethod(isNameExpr);
    }
}
